Kube proxy not touching conntrack hashsize in virtual envs

+++ tests/test-upgrade.py
@@ -33,6 +33,13 @@ class TestUpgrade(object):
         cmd = "sudo snap install microk8s --classic --channel={}".format(upgrade_from).split()
         check_call(cmd)
         wait_for_installation()
+        if in_virtual_env():
+            # In some setups (eg LXC on GCE) the hashsize nf_conntrack file under
+            # sys is marked as rw but any update on it is failing causing kube-proxy
+            # to fail.
+            here = os.path.dirname(os.path.abspath(__file__))
+            apply_patch = os.path.join(here, "patch-kube-proxy.sh")
+            check_call("sudo {}".format(apply_patch).split())
 
         # Run through the validators and
         # select those that were valid for the original snap
@@ -115,8 +122,18 @@ class TestUpgrade(object):
             print("Testing {}".format(test))
             validation()
 
-        # On lxc umount docker overlay is not permitted.
-        try:
-            check_call("sudo grep -E (lxc|hypervisor) /proc/1/environ /proc/cpuinfo".split())
-        except CalledProcessError:
+        if not in_virtual_env():
+            # On lxc umount docker overlay is not permitted.
             check_call("sudo snap remove microk8s".split())
+
+
+def in_virtual_env():
+    '''
+    Returns: True if the deployment is in a VM/container.
+
+    '''
+    try:
+        check_call("sudo grep -E (lxc|hypervisor) /proc/1/environ /proc/cpuinfo".split())
+        return True
+    except CalledProcessError:
+        return False
